requirements.txt
asgiref==3.6.0
bleach==6.0.0
Django==4.1.6
django-cors-headers==3.13.0
django-grappelli==3.0.4
django-summernote==0.8.20.0
djangorestframework==3.14.0
Pillow==9.4.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
webencodings==0.5.1
